Adding a file containing the learning database and updating the user interface
The latest version of the application has a significant update to the user interface. Replaced the hand-drawn drawing area and buttons with the pygame GUI, allowing new features to be added.

Now the user can enter a numerical value using a dedicated text field. A "Confirm" button has been added, which allows you to confirm the entered number. Additionally, the "Clear" button allows you to quickly clear the screen.

For better interaction with the user, an information label has been added that displays messages related to the entered text and number confirmation.

A new function has also been introduced to save information about white pixels to a CSV file. Each attempt to enter a number generates a new line in the data file containing the attempt number, the value entered by the user, and information about the white pixels in the grid.